created folder for routes, separated authentication and db connection into modules
